Fixed Updater
The autoupdater now works perfectly! The MbEssentials.jar file is
downloaded to plugins/mbes/data, and then move and deleted when the
server shuts down.

However, we will have to put an apology in the forum post, as the
current released version of MbEssentials has the incorrect updater code
in it. Therefore, there servers will crash when they try to update. We
will just have to tell them to ignore the crash, and just force-close
the server! :D